typo in molecule.h, changed data type of ioff in wavefunction.h to size_t, more symmetry stuff
git-svn-id: https://sirius.chem.vt.edu/svn/psi4/trunk@723 6a214465-e6f0-41bc-91e6-de896bb96b5e

+++ src/lib/libmints/wavefunction.h
@@ -8,7 +8,7 @@
 #include <psi4-dec.h>
 
 #define MAX_IOFF 30000
-extern int ioff[MAX_IOFF];
+extern size_t ioff[MAX_IOFF];
 
 #define MAX_DF 500
 extern double df[MAX_DF];
